+++ particle_tracking_manager/config_the_manager.py
@@ -6,13 +6,14 @@ import logging
 from datetime import datetime, timedelta
 from enum import Enum
 from os import PathLike
-from typing import Annotated, Self
+from typing import Annotated
 
 # Third-party imports
 import pandas as pd
 
 # from geojson import GeoJSON
 from pydantic import BaseModel, Field, computed_field, model_validator
+from typing_extensions import Self
 
 # Local imports
 from .config_ocean_model import (